src/bot_app/requirements.txt
python-telegram-bot==21.1
SQLAlchemy==2.0.15
asyncpg==0.30.0
python-dotenv==0.19.0
Flask-Login==0.6.2
werkzeug==2.3.7
pytz==2024.2